MethodHandles dispatch methods
LLVM does not, in general, allow us the facility of building up a function call in an ad-hoc basis.  While libffi could be used to implement this kind of thing, there is no support in that library for the various fast calling conventions used and supported by LLVM.

Instead, what we can do is to transform `invoke*` calls from signature-polymorphic calls by creating a structure type for each call site type.  Then we can dispatch to methods using the structure pointer argument to refer to the input arguments.

We present two forms of this operation here.  The first form allows the caller to specify the class of each argument, along with a separate pointer for each argument value.  This allows for implicit conversions in the generated dispatch method.

The second form requires that the caller and callee agree on the argument structure shape, and thus only requires an argument which points to the structure, and an argument containing a pointer into which the result should be stored.

The next step would be to create dispatch helper functions for each method handle direct target, and to implement the dispatch behavior on the various method handle types.