sidestep gap in resolution of native methods with patch type parameters
We don't currently transform method descriptors to remove patching
types when defining a class.  Therefore the descriptors of the
methods in a $_native type have to exactly match the descriptor of
the method they are implementing (effectively a $_native class is
short hand for a Patch class with a @Replace annotation on every method,
and the descriptors also have to match for @Replace to work).