better error reporting from System.initPhase2
Drop the JDK code that catches and prints any exception raised by
ModuleBootstrap.boot and just allow error to propagate back to the
interpreter.  This allows us to report the real error, instead of
reporting an error caused by trying to use native IO operations to
write an error message to an interpreted System.out.