patch out a check against System.out which isn't needed for qbicc
The JDK code was checking against System.out being null as a way of
detecting whether it was too early in bootstrapping to allow logging
to be used. This check doesn't make sense for qbicc and since we moved
System.out to a RuntimeAspect the check was now blocking build-time
initialization of the LogManager.